File name changes, bug fix in reviewList
