+++ client/src/components/App.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import QandA from './QandA.jsx';
 import AddAnswerModal from './AddAnswerModal.jsx';
-import RatingsReviews from './Ratings&Reviews/RatingsReviews.jsx';
+import RatingsReviews from './RatingsAndReviews/RatingsReviews.jsx';
 import Product from './Product.jsx';
 import appContext from '../contexts/index.js';
